♻️ refactor : Changing folder structure
- Changed the Login and Navbar components to match the convention

Related issue: #47
