+++ src/components/navbar/index.ts
@@ -1,0 +1,3 @@
+import Navbar from './Navbar';
+
+export {Navbar};
